fixed an issue with a wrong data type

+++ PlebBot/Modules/LastFm/LastFm.cs
@@ -77,7 +77,7 @@ namespace PlebBot.Modules
                         else
                         {
                             string[] columns = {"DiscordId", "LastFm"};
-                            object[] values = {Context.User.Id, username};
+                            object[] values = {(long)Context.User.Id, username};
                             await userRepo.Add(columns, values);
 
                             await Response.Success(
